SLING-5504 - Reduce memory footprint of stored recording data
Array size would be in int so avoid unnecessary cast

git-svn-id: https://svn.apache.org/repos/asf/sling/trunk/contrib/extensions/tracer@1729536 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/sling/tracer/internal/TracerLogServlet.java
@@ -80,7 +80,7 @@ class TracerLogServlet extends SimpleWebConsolePlugin implements TraceLogRecorde
                 .weigher(new Weigher<String, JSONRecording>() {
                     @Override
                     public int weigh(@Nonnull  String key, @Nonnull JSONRecording value) {
-                        return (int)value.size();
+                        return value.size();
                     }
                 })
                 .expireAfterAccess(cacheDurationInSecs, TimeUnit.SECONDS)
